Develop upstream sync 190826

+++ tensorflow/contrib/cloud/kernels/gcs_config_ops.cc
@@ -179,13 +179,13 @@ class GcsBlockCacheOpKernel : public OpKernel {
     RetryingGcsFileSystem* gcs = nullptr;
     OP_REQUIRES_OK(ctx, RetrieveGcsFs(ctx, &gcs));
 
-    size_t max_cache_size, block_size, max_staleness;
-    OP_REQUIRES_OK(ctx, ParseScalarArgument<size_t>(ctx, "max_cache_size",
+    uint64 max_cache_size, block_size, max_staleness;
+    OP_REQUIRES_OK(ctx, ParseScalarArgument<uint64>(ctx, "max_cache_size",
                                                     &max_cache_size));
     OP_REQUIRES_OK(ctx,
-                   ParseScalarArgument<size_t>(ctx, "block_size", &block_size));
+                   ParseScalarArgument<uint64>(ctx, "block_size", &block_size));
     OP_REQUIRES_OK(
-        ctx, ParseScalarArgument<size_t>(ctx, "max_staleness", &max_staleness));
+        ctx, ParseScalarArgument<uint64>(ctx, "max_staleness", &max_staleness));
 
     if (gcs->underlying()->block_size() == block_size &&
         gcs->underlying()->max_bytes() == max_cache_size &&
